Adding android support in pipelines (#22)
* Compilation warning fixes with clang android cross compilation
* Separation of Runtime-Triplet and Host-Triplet during testing
* Added additional combinations of Runtime-Triplets and Host-Triplets
* Cleaned up the vcpkg port a bit 
* Added script to auto download the android sdk

+++ include/stencil/Database2.h
@@ -655,7 +655,7 @@ template <> struct PageForRecord<0> : public PageForRecordInterface
         assert(typeId > 0xff);
         auto logRecordSize = typeId & 0xff;
         assert(logRecordSize < 12);
-        _recordSize = uint64_t{1} << logRecordSize;
+        _recordSize = size_t{1} << logRecordSize;
         switch (logRecordSize)
         {
         default:
